refactor: split mls state and channel endpoint
